fix: function typo chore: move atp agent generators
