+++ packages/bridge/src/adapters/_client/bluesky/_router.ts
@@ -15,8 +15,8 @@ import { AppAtpSessionData } from '../../../types/atproto.js';
 import BlueskyFeedRouter from './feeds.js';
 import { PostMutatorRoute } from '../_router/routes/post.js';
 import { UserRoute } from '../_router/routes/user.js';
-import { getXrpcAgent } from '../_router/_api.js';
 import { KNOWN_SOFTWARE } from '../../../data/driver.js';
+import { getXrpcAgent } from '../../../utils/atproto.js';
 
 export type AtprotoClientCreateDTO = AppAtpSessionData;
