+++ packages/bridge/src/adapters/_client/bluesky/feeds.ts
@@ -1,4 +1,3 @@
-import { getXrpcAgent } from '../_router/_api.js';
 import { AppAtpSessionData } from '../../../types/atproto.js';
 import { SavedFeedsPrefV2 } from '@atproto/api/dist/client/types/app/bsky/actor/defs.js';
 import { RandomUtil } from '../../../utils/random.js';
@@ -13,6 +12,7 @@ import {
 import { Err, Ok, ResultErr, ResultOk } from '../../../utils/result.js';
 import { ApiErrorCode } from '../../../types/result.types.js';
 import type { ApiAsyncResult } from '../../../utils/api-result.js';
+import { getXrpcAgent } from '../../../utils/atproto.js';
 
 type SubscriptionUpdateResult = Promise<{
 	success: boolean;
